Split POST /applications/{applicationId}/submission api endpoint
