+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/integration/ApplicationTest.kt
@@ -39,7 +39,6 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.RoshRisksEnvel
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.SentenceTypeOption
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ServiceName
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.SubmitApprovedPremisesApplication
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.SubmitTemporaryAccommodationApplication
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.UnknownPerson
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.UpdateApplicationType
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.UpdateApprovedPremisesApplication
@@ -71,7 +70,6 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.AssessmentEnt
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.AutoAllocationDay
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.Cas1CruManagementAreaEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.DomainEventType
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.TemporaryAccommodationAssessmentEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserQualification
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserRole
@@ -1783,217 +1781,6 @@ class ApplicationTest : IntegrationTestBase() {
       }
     }
   }
-
-  @Nested
-  inner class Cas3SubmitApplication {
-
-    @Test
-    fun `Submit Temporary Accommodation application returns 200`() {
-      givenAUser { submittingUser, jwt ->
-        givenAUser { _, _ ->
-          givenAnOffender { offenderDetails, _ ->
-            val applicationId = UUID.fromString("22ceda56-98b2-411d-91cc-ace0ab8be872")
-            val offenderName = "${offenderDetails.firstName} ${offenderDetails.surname}"
-
-            val applicationSchema = temporaryAccommodationApplicationJsonSchemaEntityFactory.produceAndPersist {
-              withAddedAt(OffsetDateTime.now())
-              withId(UUID.randomUUID())
-              withSchema(schemaText())
-            }
-
-            temporaryAccommodationApplicationEntityFactory.produceAndPersist {
-              withCrn(offenderDetails.otherIds.crn)
-              withId(applicationId)
-              withApplicationSchema(applicationSchema)
-              withCreatedByUser(submittingUser)
-              withProbationRegion(submittingUser.probationRegion)
-              withName(offenderName)
-              withData(
-                """
-                {}
-              """,
-              )
-            }
-
-            webTestClient.post()
-              .uri("/applications/$applicationId/submission")
-              .header("Authorization", "Bearer $jwt")
-              .header("X-Service-Name", ServiceName.temporaryAccommodation.value)
-              .bodyValue(
-                SubmitTemporaryAccommodationApplication(
-                  translatedDocument = {},
-                  type = "CAS3",
-                  arrivalDate = LocalDate.now(),
-                  summaryData = object {
-                    val num = 50
-                    val text = "Hello world!"
-                  },
-                ),
-              )
-              .exchange()
-              .expectStatus()
-              .isOk
-
-            val persistedApplication = temporaryAccommodationApplicationRepository.findByIdOrNull(applicationId)!!
-            val persistedAssessment = persistedApplication.getLatestAssessment() as TemporaryAccommodationAssessmentEntity
-
-            assertThat(persistedAssessment.summaryData).isEqualTo("{\"num\":50,\"text\":\"Hello world!\"}")
-            assertThat(persistedApplication.name).isEqualTo(offenderName)
-          }
-        }
-      }
-    }
-
-    @Test
-    fun `Submit Temporary Accommodation application returns 200 with optional elements in the request`() {
-      givenAUser { submittingUser, jwt ->
-        givenAUser { _, _ ->
-          givenAnOffender { offenderDetails, _ ->
-            val applicationId = UUID.fromString("22ceda56-98b2-411d-91cc-ace0ab8be872")
-
-            val applicationSchema = temporaryAccommodationApplicationJsonSchemaEntityFactory.produceAndPersist {
-              withAddedAt(OffsetDateTime.now())
-              withId(UUID.randomUUID())
-              withSchema(schemaText())
-            }
-
-            temporaryAccommodationApplicationEntityFactory.produceAndPersist {
-              withCrn(offenderDetails.otherIds.crn)
-              withId(applicationId)
-              withApplicationSchema(applicationSchema)
-              withCreatedByUser(submittingUser)
-              withProbationRegion(submittingUser.probationRegion)
-              withPdu("Probation Delivery Unit Test")
-              withHasHistoryOfSexualOffence(true)
-              withIsConcerningSexualBehaviour(true)
-              withIsConcerningArsonBehaviour(true)
-              withData(
-                """
-              {}
-            """,
-              )
-            }
-
-            val pdu = probationDeliveryUnitFactory.produceAndPersist {
-              withProbationRegion(
-                probationRegionEntityFactory.produceAndPersist {
-                  withId(submittingUser.probationRegion.id)
-                },
-              )
-            }
-
-            webTestClient.post()
-              .uri("/applications/$applicationId/submission")
-              .header("Authorization", "Bearer $jwt")
-              .header("X-Service-Name", ServiceName.temporaryAccommodation.value)
-              .bodyValue(
-                SubmitTemporaryAccommodationApplication(
-                  translatedDocument = {},
-                  type = "CAS3",
-                  arrivalDate = LocalDate.now(),
-                  summaryData = object {
-                    val num = 50
-                    val text = "Hello world!"
-                  },
-                  personReleaseDate = LocalDate.now(),
-                  pdu = "Probation Delivery Unit Test",
-                  isHistoryOfSexualOffence = true,
-                  isConcerningSexualBehaviour = true,
-                  isConcerningArsonBehaviour = true,
-                  dutyToReferOutcome = "Accepted – Prevention/ Relief Duty",
-                  prisonReleaseTypes = listOf(
-                    "Parole",
-                    "CRD licence",
-                  ),
-                  probationDeliveryUnitId = pdu.id,
-                ),
-              )
-              .exchange()
-              .expectStatus()
-              .isOk
-
-            val persistedApplication = temporaryAccommodationApplicationRepository.findByIdOrNull(applicationId)!!
-            val persistedAssessment =
-              persistedApplication.getLatestAssessment() as TemporaryAccommodationAssessmentEntity
-
-            assertThat(persistedAssessment.summaryData).isEqualTo("{\"num\":50,\"text\":\"Hello world!\"}")
-            assertThat(persistedApplication.personReleaseDate).isEqualTo(LocalDate.now())
-            assertThat(persistedApplication.dutyToReferOutcome).isEqualTo("Accepted – Prevention/ Relief Duty")
-            assertThat(persistedApplication.prisonReleaseTypes).isEqualTo("Parole,CRD licence")
-            assertThat(persistedApplication.probationDeliveryUnit!!.id).isEqualTo(pdu.id)
-            assertThat(persistedApplication.probationDeliveryUnit!!.name).isEqualTo(pdu.name)
-          }
-        }
-      }
-    }
-
-    @Test
-    fun `Submit Temporary Accommodation application returns 400 when the application was deleted`() {
-      givenAUser { submittingUser, jwt ->
-        givenAUser { _, _ ->
-          givenAnOffender { offenderDetails, _ ->
-            val applicationId = UUID.fromString("22ceda56-98b2-411d-91cc-ace0ab8be872")
-            val offenderName = "${offenderDetails.firstName} ${offenderDetails.surname}"
-
-            val applicationSchema = temporaryAccommodationApplicationJsonSchemaEntityFactory.produceAndPersist {
-              withAddedAt(OffsetDateTime.now())
-              withId(UUID.randomUUID())
-              withSchema(schemaText())
-            }
-
-            temporaryAccommodationApplicationEntityFactory.produceAndPersist {
-              withCrn(offenderDetails.otherIds.crn)
-              withId(applicationId)
-              withApplicationSchema(applicationSchema)
-              withCreatedByUser(submittingUser)
-              withProbationRegion(submittingUser.probationRegion)
-              withName(offenderName)
-              withDeletedAt(OffsetDateTime.now().minusDays(32))
-              withData(
-                """
-                {}
-              """,
-              )
-            }
-
-            webTestClient.post()
-              .uri("/applications/$applicationId/submission")
-              .header("Authorization", "Bearer $jwt")
-              .header("X-Service-Name", ServiceName.temporaryAccommodation.value)
-              .bodyValue(
-                SubmitTemporaryAccommodationApplication(
-                  translatedDocument = {},
-                  type = "CAS3",
-                  arrivalDate = LocalDate.now(),
-                  summaryData = object {
-                    val num = 50
-                    val text = "Hello world!"
-                  },
-                ),
-              )
-              .exchange()
-              .expectStatus()
-              .isBadRequest
-              .expectBody()
-              .jsonPath("$.status").isEqualTo("400")
-              .jsonPath("$.detail").isEqualTo("This application has already been deleted")
-          }
-        }
-      }
-    }
-  }
-
-  private fun schemaText(): String = """
-              {
-                "${"\$schema"}": "https://json-schema.org/draft/2020-12/schema",
-                "${"\$id"}": "https://example.com/product.schema.json",
-                "title": "Thing",
-                "description": "A thing",
-                "type": "object",
-                "properties": {},
-                "required": []
-              }
-            """
 
   @Nested
   inner class GetAssessmentForApplication {
